env.txt: Set bootdelay to 0
If a serial cable is not attached, the missing pull-up resistor on the
RX pin can cause jitter. This can be interpreted as a key press by
uboot.

Setting bootdelay to 0 prevents the board getting stuck at the uboot
prompt at boot, but it will also be impossible to stop the boot with a
key press.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>